backend/requirements.txt
# ─── NEXUS RAG — Python Dependencies ─────────────────────────────────────────

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# LLM
groq>=0.9.0

# Vector store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# PDF parsing
pypdf>=4.2.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment
python-dotenv>=1.0.0

# Utilities
numpy>=1.26.0
